requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
fastapi-users[sqlalchemy]
databases
fastapi_localization
pytest
psycopg2
celery
fastapi-cache2[redis]
aioredis
flower
password-validator